hermes-adapter/requirements.txt
dingtalk-stream>=0.20
httpx>=0.24
pyyaml>=6.0
